Implemention of Python List using C
IMPLEMENTATION OF PYTHON LIST USING LINKED LIST IN C

OUTPUT

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
1
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 2

Enter the element to be inserted :AS
PS C:\kalpavriksha> gcc pythonList.c
PS C:\kalpavriksha> gcc pythonList.c 
PS C:\kalpavriksha> ./a.exe

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
1
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 2

Enter the element to be inserted :DS
PS C:\kalpavriksha> gcc pythonList.c
PS C:\kalpavriksha> ./a.exe

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
1
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 0

Enter the element to be inserted :23
23
inserted at begining
Do you want to continue? (Enter any number for Yes or 0 for No) 5

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
2
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 1

Enter the element to be inserted :4.4
4.400000
 inserted at end
Do you want to continue? (Enter any number for Yes or 0 for No) 5

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
8
23 ->4.400000 ->NULL
Do you want to continue? (Enter any number for Yes or 0 for No)
PS C:\kalpavriksha> gcc pythonList.c
PS C:\kalpavriksha> ./a.exe

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
1
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 2

Enter the element to be inserted :sdfa
sdfa
inserted at begining
Do you want to continue? (Enter any number for Yes or 0 for No) 5

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
2
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 1

Enter the element to be inserted :5.5
5.500000
 inserted at end
Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
3
Enter the data type (0 for INT, 1 for FLOAT, 2 for STRING): 0

Enter the element to be inserted :45

Enter the position2
45
 inserted at position 2
Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
8
sdfa ->45 ->5.500000 ->NULL
Do you want to continue? (Enter any number for Yes or 0 for No)