Implementation of LinkedList in C
Output:
Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
1

Enter the element to insert at begining :45

Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
2

Enter the element to insert at end :24

24 inserted at end
Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
3

Enter the element to insert at position2

Enter the position2

2 inserted at position 2
Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
8
45->2->24->NULL
Do you want to continue? (Enter any number for Yes or 0 for No)
Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
8
45->2->24->NULL
Do you want to continue? (Enter any number for Yes or 0 for No) 1

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
6

Enter the position at which the element has to be deleted2

2 deleted from position 2
Do you want to continue? (Enter any number for Yes or 0 for No) 1 

Enter the operation you want to perform in Linked list:
1.Insert at Begining
2. Insert at end
3. Insert at the Position
4. Delete from Begining
5.Delete from end
6.Delete from a Position
7.size
8. display
9. Exit
8
45->24->NULL
Do you want to continue? (Enter any number for Yes or 0 for No)